testing binding of Poco::File::exists() and Poco::File::getSize() work.
cmake builds libfoundation.so, but copying to foundation.so enables lua to load it properly.
there is a small concern about the library name embedded into the shared object that might cause problems later by just renaming it,so it would probably be wise to set the CMAKE prefix properly.